+++ release-notes.md
@@ -1,6 +1,6 @@
 ## Change Notes for {ver}:
 
-* 
+* Clean up error reporting for unknown code systems
 
 ## Conformance Notes: